ENG-56603 - Auth0 Workflow Preservation
In order to support authentication via keycloak in environments without
3rd party cookies, it is essential to initialize the keycloak adapter
BEFORE angular takes control of the application's URL.  However, this
preemptive initialization also causes keycloak to override auth0
workflows because both are OIDC implementations that share a subset of
query parameters.  This change attempts to differentiate between
keycloak and auth0 workflows so that keycloak can still be initialized
preemptively, but NOT when it is inappropriate.