ENG-0000 - Target AIMS Global Stack/Token Info via cookie
Updates all AIMS endpoints to use the global rather than local stack.
According to James, this should have always been the case; the local
stack redirects internally to the global one, but it is less efficient.

Updated the token_info endpoint to omit any explicit token when
running as a plugin.